backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1
alembic==1.13.1
celery==5.4.0
redis==5.0.6
google-generativeai==0.7.2
httpx==0.27.0
